[torch-frontend] update torch to 2.4.1
