+++ frontends/torch-frontend/torch-cpu-requirements.txt
@@ -1,4 +1,4 @@
 --extra-index-url https://download.pytorch.org/whl/cpu
 --pre
-torch==2.1.0+cpu
-torchvision==0.16.0+cpu
+torch==2.4.1+cpu
+torchvision==0.19.1+cpu
